pyproject.toml
[tool.poetry]
name = "delta"
version = "0.1.0"
description = "Data-driven Evaluation and Live Trading Algorithm"
authors = ["BrianBae <bohblue23@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.24.1"
prefect = "^2.13.0"
hvac = "^1.2.1"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.5"
orjson = "^3.9.7"
tqdm = "^4.66.1"
sentry-sdk = "^1.30.0"
pyzmq = "^25.1.1"


[tool.poetry.group.dev.dependencies]
click = "^8.1.7"
clipboard = "^0.0.4"
black = "^23.9.0"
ruff = "^0.0.287"
pre-commit = "^3.4.0"
isort = "^5.12.0"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.codespell]
exclude-file = "delta/adapter/ebest/block.py"

[tool.ruff]
line-length = 120